+++ etherman/etherman_x1.go
@@ -2,6 +2,7 @@ package etherman
 
 import (
 	"crypto/ecdsa"
+	"errors"
 	"fmt"
 	"github.com/0xPolygonHermez/zkevm-node/etherman/smartcontracts/polygonzkevm"
 	ethmanTypes "github.com/0xPolygonHermez/zkevm-node/etherman/types"
@@ -124,4 +125,34 @@ func newAuthFromKeystoreX1(path, password string, chainID uint64) (bind.Transact
 		return bind.TransactOpts{}, nil, err
 	}
 	return *auth, key.PrivateKey, nil
+}
+
+// generateMockAuth generates an authorization instance from a
+// randomly generated private key to be used to estimate gas for PoE
+// operations NOT restricted to the Trusted Sequencer
+func (etherMan *Client) generateMockAuth(sender common.Address) (bind.TransactOpts, error) {
+	privateKey, err := crypto.GenerateKey()
+	if err != nil {
+		return bind.TransactOpts{}, errors.New("failed to generate a private key to estimate L1 txs")
+	}
+	chainID := big.NewInt(0).SetUint64(etherMan.l1Cfg.L1ChainID)
+	auth, err := bind.NewKeyedTransactorWithChainID(privateKey, chainID)
+	if err != nil {
+		return bind.TransactOpts{}, errors.New("failed to generate a fake authorization to estimate L1 txs")
+	}
+
+	auth.From = sender
+	auth.Signer = func(address common.Address, tx *types.Transaction) (*types.Transaction, error) {
+		chainID := big.NewInt(0).SetUint64(etherMan.l1Cfg.L1ChainID)
+		signer := types.LatestSignerForChainID(chainID)
+		if err != nil {
+			return nil, err
+		}
+		signature, err := crypto.Sign(signer.Hash(tx).Bytes(), privateKey)
+		if err != nil {
+			return nil, err
+		}
+		return tx.WithSignature(signer, signature)
+	}
+	return *auth, nil
 }
