+++ jsonrpc/server.go
@@ -14,7 +14,6 @@ import (
 	"time"
 
 	"github.com/0xPolygonHermez/zkevm-node/jsonrpc/metrics"
-	"github.com/0xPolygonHermez/zkevm-node/jsonrpc/nacos"
 	"github.com/0xPolygonHermez/zkevm-node/jsonrpc/types"
 	"github.com/0xPolygonHermez/zkevm-node/log"
 	"github.com/didip/tollbooth/v6"
@@ -528,16 +527,4 @@ func (s *Server) combinedLog(r *http.Request, start time.Time, httpStatus, dataL
 		r.Host,
 		r.UserAgent(),
 	)
-}
-
-func (s *Server) registerNacos() {
-	// start nacos client for registering restful service
-	if s.config.Nacos.URLs != "" {
-		nacos.StartNacosClient(s.config.Nacos.URLs, s.config.Nacos.NamespaceId, s.config.Nacos.ApplicationName, s.config.Nacos.ExternalListenAddr)
-	}
-
-	// start nacos client for registering restful service
-	if s.config.NacosWs.URLs != "" {
-		nacos.StartNacosClient(s.config.NacosWs.URLs, s.config.NacosWs.NamespaceId, s.config.NacosWs.ApplicationName, s.config.NacosWs.ExternalListenAddr)
-	}
 }
